+++ test/support/postgres/test_repo.ex
@@ -28,4 +28,9 @@ if Code.ensure_loaded?(Ecto.Adapters.SQL) and Code.ensure_loaded?(Postgrex) do
 
     use Ecto.Repo, otp_app: :docket, adapter: Ecto.Adapters.Postgres
   end
+
+  defmodule Docket.Postgres.EventStoreTestRepo do
+    @moduledoc false
+    use Ecto.Repo, otp_app: :docket, adapter: Ecto.Adapters.Postgres
+  end
 end